Implement Parsing.
This works with a (very small) library ("Parse") that... somewhat looks
like some kind of parser combinator library (only this one sucks).

It has enough to parse this lisp dialect, though, so it's all good.

Also refactored the lexer to consider the operators as "id"s (which they
should be, seeing as, well, it's lisp).

Added the parsing case to the main file.

TODO now: add tests (tons of tests)